now when you are about to delete a note you are asked to confirm

+++ script.js
@@ -66,6 +66,8 @@ function showMenu(elem){
 }
 
 function deleteNote(noteId){
+    let confirmDel = confirm("Are you sure you want to delete this note?");
+    if(!confirmDel) return;
     notes.splice(noteId, 1);
     localStorage.setItem("notes",JSON.stringify(notes));
     showNotes();
